+++ source/Flotsam.h
@@ -49,7 +49,7 @@ public:
 	void Place(const Ship &source);
 	// Place this flotsam with its starting position at the specified bay of the source ship,
 	// instead of the center of the ship.
-	void Place(const Ship &source, int bayIndex);
+	void Place(const Ship &source, size_t bayIndex);
 	// Place flotsam coming from something other than a ship. Optionally specify
 	// the maximum relative velocity, or the exact relative velocity as a vector.
 	void Place(const Body &source, double maxVelocity = .5);
